+++ packages/color-picker/typings/ColorPickerProps.d.ts
@@ -1,5 +1,5 @@
 /**
- * This file was generated from ColorWheel.xml
+ * This file was generated from ColorPicker.xml
  * WARNING: All changes made to this file will be overwritten
  * @author Mendix Widgets Team
  */
@@ -8,7 +8,7 @@ interface CommonProps<Style> {
     style: Style[];
 }
 
-export interface ColorWheelProps<Style> extends CommonProps<Style> {
+export interface ColorPickerProps<Style> extends CommonProps<Style> {
     color: EditableValue<string>;
     onChange?: ActionValue;
     onChangeComplete?: ActionValue;